Move fields into new module & tidy up imports
