+++ src/physics/mod.rs
@@ -1,3 +1,4 @@
 pub mod types;
 pub mod force;
 pub mod barneshut;
+pub mod field;
